pyproject.toml
[project]
name = "himitsu"
version = "0.1.0"
description = "A Simple CLI-based Password Manager written in Python"
authors = [
    {name = "Nursultan Zhantuar",email = "nursultanzhantuar@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography (>=45.0.5,<46.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
